requirements.txt
pytesseract
opencv-python
pillow

numpy
openpyxl
pytesseract
